+++ turbopack/crates/turbopack-trace-server/src/bottom_up.rs
@@ -1,10 +1,11 @@
 use std::{env, sync::Arc};
 
-use hashbrown::{Equivalent, HashMap};
+use hashbrown::HashMap;
 
 use crate::{
     span::{SpanBottomUp, SpanIndex},
     span_ref::SpanRef,
+    string_tuple_ref::StringTupleRef,
 };
 
 pub struct SpanBottomUpBuilder {
@@ -32,33 +33,6 @@ impl SpanBottomUpBuilder {
                 .map(|child| Arc::new(child.build()))
                 .collect(),
         )
-    }
-}
-
-#[derive(Hash)]
-struct StringTupleRef<'a>(&'a str, &'a str);
-
-impl<'a> Equivalent<(String, String)> for StringTupleRef<'a> {
-    fn equivalent(&self, other: &(String, String)) -> bool {
-        self.0 == other.0 && self.1 == other.1
-    }
-}
-
-#[cfg(test)]
-mod string_tuple_ref_tests {
-    use std::hash::RandomState;
-
-    use super::*;
-
-    #[test]
-    fn test_string_tuple_ref_hash() {
-        use std::hash::BuildHasher;
-
-        let s = RandomState::new();
-        assert_eq!(
-            s.hash_one(&StringTupleRef("abc", "def")),
-            s.hash_one(&("abc".to_string(), "def".to_string()))
-        );
     }
 }
