+++ turbopack/crates/turbopack-trace-server/src/main.rs
@@ -18,6 +18,7 @@ mod span_graph_ref;
 mod span_ref;
 mod store;
 mod store_container;
+mod string_tuple_ref;
 mod timestamp;
 mod u64_empty_string;
 mod u64_string;
