+++ turbopack/crates/turbopack-trace-server/src/span_bottom_up_ref.rs
@@ -6,9 +6,9 @@ use std::{
 
 use crate::{
     FxIndexMap,
-    span::{SpanBottomUp, SpanGraphEvent, SpanIndex},
+    span::{SpanBottomUp, SpanGraphEvent},
     span_graph_ref::{SpanGraphEventRef, SpanGraphRef, event_map_to_list},
-    span_ref::SpanRef,
+    span_ref::{GroupNameToDirectAndRecusiveSpans, SpanRef},
     store::{SpanId, Store},
     timestamp::Timestamp,
 };
@@ -85,8 +85,7 @@ impl<'a> SpanBottomUpRef<'a> {
                     let _ = self.first_span().graph();
                     self.first_span().extra().graph.get().unwrap().clone()
                 } else {
-                    let mut map: FxIndexMap<(&str, &str), (Vec<SpanIndex>, Vec<SpanIndex>)> =
-                        FxIndexMap::default();
+                    let mut map: GroupNameToDirectAndRecusiveSpans = FxIndexMap::default();
                     let mut queue = VecDeque::with_capacity(8);
                     for child in self.spans() {
                         let name = child.group_name();
