+++ turbopack/crates/turbopack-trace-server/src/span_graph_ref.rs
@@ -9,9 +9,9 @@ use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 use crate::{
     FxIndexMap,
     bottom_up::build_bottom_up_graph,
-    span::{SpanGraph, SpanGraphEvent, SpanIndex},
+    span::{SpanGraph, SpanGraphEvent},
     span_bottom_up_ref::SpanBottomUpRef,
-    span_ref::SpanRef,
+    span_ref::{GroupNameToDirectAndRecusiveSpans, SpanRef},
     store::{SpanId, Store},
     timestamp::Timestamp,
 };
@@ -87,8 +87,7 @@ impl<'a> SpanGraphRef<'a> {
                 self.first_span().extra().graph.get().unwrap().clone()
             } else {
                 let self_group = self.first_span().group_name();
-                let mut map: FxIndexMap<(&str, &str), (Vec<SpanIndex>, Vec<SpanIndex>)> =
-                    FxIndexMap::default();
+                let mut map: GroupNameToDirectAndRecusiveSpans = FxIndexMap::default();
                 let mut queue = VecDeque::with_capacity(8);
                 for span in self.recursive_spans() {
                     for span in span.children() {
@@ -303,9 +302,7 @@ impl<'a> SpanGraphRef<'a> {
     }
 }
 
-pub fn event_map_to_list(
-    map: FxIndexMap<(&str, &str), (Vec<SpanIndex>, Vec<SpanIndex>)>,
-) -> Vec<SpanGraphEvent> {
+pub fn event_map_to_list(map: GroupNameToDirectAndRecusiveSpans) -> Vec<SpanGraphEvent> {
     map.into_iter()
         .map(|(_, (root_spans, recursive_spans))| {
             let graph = SpanGraph {
